+++ crypto-sodium/CHANGELOG.md
@@ -6,3 +6,4 @@
 
 * `sodiumInit`
 * Random bytes generation: `Crypto.Random`
+* Key derivation: `Crypto.Key.fromPassword`